[Args] move args to main
