+++ src/interpret.rs
@@ -1,4 +1,3 @@
-use crate::args::Args;
 use crate::utils::debug::*;
 use crate::{interpreter, parser, typing};
