+++ src/lib.rs
@@ -1,4 +1,3 @@
-mod args;
 pub mod ast;
 pub mod error;
 pub mod interpreter;
