+++ src/main.rs
@@ -1,4 +1,3 @@
-mod args;
 mod ast;
 mod error;
 mod interpret;
@@ -13,8 +12,30 @@ mod utils;
 use clap::Parser;
 use colored::control::SHOULD_COLORIZE;
 
+#[derive(Parser, Debug)]
+#[command(version, about, long_about = None)]
+pub struct Args {
+    #[arg(help = "filename")]
+    pub file_name: Option<String>,
+
+    #[arg(long, help = "Print sexp tree-sitter parser")]
+    pub debug_sexp: bool,
+
+    #[arg(long, help = "Print the AST")]
+    pub debug_parser: bool,
+
+    #[arg(long, help = "Print the typed program")]
+    pub debug_typer: bool,
+
+    #[arg(long, help = "Print the interpreter")]
+    pub debug_interpreter: bool,
+
+    #[arg(long, help = "Don't colorize error")]
+    pub no_color: bool,
+}
+
 fn main() {
-    let args = args::Args::parse();
+    let args = Args::parse();
     if args.no_color {
         SHOULD_COLORIZE.set_override(false)
     }
